+++ client/src/app/videos/+video-watch/video-watch.component.ts
@@ -125,17 +125,17 @@ export class VideoWatchComponent implements OnInit, OnDestroy {
       new Hotkey('shift+l', (event: KeyboardEvent): boolean => {
         this.setLike()
         return false
-      }, undefined, 'Like the video'),
+      }, undefined, this.i18n('Like the video')),
       new Hotkey('shift+d', (event: KeyboardEvent): boolean => {
         this.setDislike()
         return false
-      }, undefined, 'Dislike the video'),
+      }, undefined, this.i18n('Dislike the video')),
       new Hotkey('shift+s', (event: KeyboardEvent): boolean => {
         this.subscribeButton.subscribed ?
           this.subscribeButton.unsubscribe() :
           this.subscribeButton.subscribe()
         return false
-      }, undefined, 'Subscribe to the account')
+      }, undefined, this.i18n('Subscribe to the account'))
     ]
     if (this.isUserLoggedIn()) this.hotkeysService.add(this.hotkeys)
   }
@@ -369,13 +369,17 @@ export class VideoWatchComponent implements OnInit, OnDestroy {
         )
   }
 
-  private async onVideoFetched (video: VideoDetails, videoCaptions: VideoCaption[], startTime = 0) {
+  private async onVideoFetched (video: VideoDetails, videoCaptions: VideoCaption[], startTimeFromUrl: number) {
     this.video = video
 
     // Re init attributes
     this.descriptionLoading = false
     this.completeDescriptionShown = false
     this.remoteServerDown = false
+
+    let startTime = startTimeFromUrl || (this.video.userHistory ? this.video.userHistory.currentTime : 0)
+    // Don't start the video if we are at the end
+    if (this.video.duration - startTime <= 1) startTime = 0
 
     if (this.video.isVideoNSFWForUser(this.user, this.serverService.getConfig())) {
       const res = await this.confirmService.confirm(
@@ -414,7 +418,12 @@ export class VideoWatchComponent implements OnInit, OnDestroy {
       poster: this.video.previewUrl,
       startTime,
       theaterMode: true,
-      language: this.localeId
+      language: this.localeId,
+
+      userWatching: this.user ? {
+        url: this.videoService.getUserWatchingVideoUrl(this.video.uuid),
+        authorizationHeader: this.authService.getRequestHeaderValue()
+      } : undefined
     })
 
     if (this.videojsLocaleLoaded === false) {